requirements.txt
pypdf
langchain
torch
accelerate
bitsandbytes
transformers
sentence_transformers
faiss_cpu
streamlit
typing-inspect==0.8.0
typing_extensions==4.5.0
pydantic 
ctransformers
numpy